Inline config-only Enums
